+++ src/main/java/com/micro/namecard/EsConfiguration.java
@@ -1,6 +1,7 @@
-package com.micro.search;
+package com.micro.namecard;
 
 import org.elasticsearch.client.RestHighLevelClient;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.elasticsearch.client.ClientConfiguration;
@@ -10,6 +11,9 @@ import org.springframework.data.elasticsearch.config.AbstractElasticsearchConfig
 @Configuration
 public class EsConfiguration extends AbstractElasticsearchConfiguration {
 
+    @Value(value="${es.bootstrap.server}")
+    private String elasticBootServer;
+
     /**
      * Elastic Search Client Configuration and Creation.
      */
@@ -17,7 +21,7 @@ public class EsConfiguration extends AbstractElasticsearchConfiguration {
     @Bean
     public RestHighLevelClient elasticsearchClient() {
         ClientConfiguration clientConfiguration = ClientConfiguration.builder()
-                .connectedTo("localhost:9200")
+                .connectedTo(elasticBootServer)
                 .build();
 
         return RestClients.create(clientConfiguration).rest();
