+++ src/main/java/com/micro/namecard/SearchApplication.java
@@ -1,4 +1,4 @@
-package com.micro.search;
+package com.micro.namecard;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
