+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
-
+server.port=8080
+es.bootstrap.server=localhost:9200
